refactor: utility type 적용
