+++ src/components/home/walking/WalkingDongilItem.tsx
@@ -1,7 +1,5 @@
 import { Link, useNavigate } from 'react-router-dom';
 import styled from 'styled-components';
-import { TItemName } from '@lib/types/TItemName';
-import { TInterestRate } from '@lib/types/IInterestRate';
 import renderItemIllust from '@lib/utils/render/renderItemIllust';
 import { ReactComponent as Failed } from '@assets/icons/failed.svg';
 import { ReactComponent as Arrow } from '@assets/icons/arrow-walking.svg';
@@ -10,7 +8,6 @@ import CommonSheet from '@components/common/bottomSheets/CommonSheet';
 import DongilFailed from '@components/common/bottomSheets/sheetContents/DongilFailed';
 import DeleteCheck from '@components/common/bottomSheets/sheetContents/DeleteCheck';
 import SheetCompleted from '@components/common/bottomSheets/sheetContents/SheetCompleted';
-import { TDongilStatus } from '@lib/types/TDongilStatus';
 import useAxiosPrivate from '@lib/hooks/auth/useAxiosPrivate';
 import { useState } from 'react';
 import { TFetchStatus } from '@lib/types/TFetchStatus';
@@ -19,14 +16,13 @@ import {
   deleteClientSideWalkingDongilById,
   deleteWalkingDongil,
 } from '@store/slices/walkingDongilsSlice';
+import { IDongil } from '@lib/types/IDongil';
 
-interface WalkingDongilItemProps {
-  itemName: TItemName;
-  title: string;
-  id: number;
-  interestRate: TInterestRate;
-  challengeStatus: TDongilStatus;
-}
+interface WalkingDongilItemProps
+  extends Pick<
+    IDongil,
+    'itemName' | 'title' | 'id' | 'interestRate' | 'challengeStatus'
+  > {}
 
 function WalkingDongilItem({
   itemName,
